Static 2.

+++ src/test/get_model_metadata_signature_test.cpp
@@ -88,7 +88,7 @@ TEST_F(GetModelMetadataSignature, ConvertCorrectTensorShape) {
             return false;
         }
         for (int i = 0; i < actual.dim_size(); i++) {
-            if (actual.dim(i).size() != expected[i]) {
+            if (static_cast<unsigned int>(actual.dim(i).size()) != expected[i]) {
                 return false;
             }
         }
